% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-lazy.R
\name{add_index}
\alias{add_index}
\title{xxx xxx xxx}
\usage{
add_index(tibble, name_index = "index")
}
\arguments{
\item{tibble}{xxx xxx xxx}

\item{name_index}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
