% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{as_any_symbol}
\alias{as_any_symbol}
\title{xxx xxx xxx}
\usage{
as_any_symbol(x)
}
\arguments{
\item{x}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
