% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-viz.R
\name{bookdown_template}
\alias{bookdown_template}
\title{Create a bookdown template.}
\usage{
bookdown_template(bookdown_path, overwrite = FALSE)
}
\arguments{
\item{bookdown_path}{A character string identifying the folder path where
the bookdown will be generated.}

\item{overwrite}{whether to overwrite existing files. FALSE by default.}
}
\value{
A folder containing all files (Rmd, yml, css) to generate the bookdown.
}
\description{
This helper function creates a template for a bookdown.
}
\examples{
{

bookdown_path = tempdir()
bookdown_template(bookdown_path, overwrite = TRUE)

}

}
\seealso{
\code{\link[=bookdown_render]{bookdown_render()}},\code{\link[=bookdown_open]{bookdown_open()}}
}
