% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-functions_template.R
\name{cast_error_fargs}
\alias{cast_error_fargs}
\title{xxx xxx xxx}
\usage{
cast_error_fargs(...)
}
\arguments{
\item{...}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
