% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-lazy.R
\name{collect_roxygen}
\alias{collect_roxygen}
\title{xxx xxx xxx}
\usage{
collect_roxygen(folder_r = "R")
}
\arguments{
\item{folder_r}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
