% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{template_visual_report}
\alias{plot_main_word}
\alias{plot_histogram}
\alias{plot_box}
\alias{plot_bar}
\alias{plot_date}
\alias{plot_density}
\alias{plot_pie}
\alias{plot_pie_valid_value}
\alias{summary_text}
\alias{summary_numerical}
\alias{summary_category}
\title{Deprecated functions}
\usage{
template_visual_report(...)

plot_main_word(...)

plot_histogram(...)

plot_box(...)

plot_bar(...)

plot_date(...)

plot_density(...)

plot_pie(...)

plot_pie_valid_value(...)

summary_text(...)

summary_numerical(...)

summary_category(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link[=bookdown_template]{bookdown_template()}} instead of \code{template_visual_report()}.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

For any fabR::plot_xxx() or summary_xxx() use
\code{\link[madshapR:variable_visualize]{madshapR::variable_visualize()}} instead.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

For any fabR::plot_xxx() or summary_xxx() use
\code{\link[madshapR:variable_visualize]{madshapR::variable_visualize()}} instead.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

For any fabR::plot_xxx() or summary_xxx() use
\code{\link[madshapR:variable_visualize]{madshapR::variable_visualize()}} instead.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

For any fabR::plot_xxx() or summary_xxx() use
\code{\link[madshapR:variable_visualize]{madshapR::variable_visualize()}} instead.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

For any fabR::plot_xxx() or summary_xxx() use
\code{\link[madshapR:variable_visualize]{madshapR::variable_visualize()}} instead.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

For any fabR::plot_xxx() or summary_xxx() use
\code{\link[madshapR:variable_visualize]{madshapR::variable_visualize()}} instead.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

For any fabR::plot_xxx() or summary_xxx() use
\code{\link[madshapR:variable_visualize]{madshapR::variable_visualize()}} instead.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

For any fabR::plot_xxx() or summary_xxx() use
\code{\link[madshapR:variable_visualize]{madshapR::variable_visualize()}} instead.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

For any fabR::plot_xxx() or summary_xxx() use
\code{\link[madshapR:variable_visualize]{madshapR::variable_visualize()}} instead.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

For any fabR::plot_xxx() or summary_xxx() use
\code{\link[madshapR:variable_visualize]{madshapR::variable_visualize()}} instead.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

For any fabR::plot_xxx() or summary_xxx() use
\code{\link[madshapR:variable_visualize]{madshapR::variable_visualize()}} instead.
}
\keyword{internal}
