% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{fabR_help}
\alias{fabR_help}
\title{Call the help center for full documentation}
\usage{
fabR_help()
}
\value{
Nothing to be returned. The function opens a web package.
}
\description{
This feature is a direct call of the documentation in the repository hosting
the package. The user accesses the description of the latest version of the
package, the vignettes, and the list of functions.
}
\examples{
{

# call the help center!
fabR_help()

}

}
