% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-check_table_values.R
\name{get_all_na_rows2}
\alias{get_all_na_rows2}
\title{Extract observations (rows) that have all NA values in a tibble (no ID)}
\usage{
get_all_na_rows2(tbl)
}
\arguments{
\item{tbl}{R object(dataframe or tibble) of the input tibble}
}
\value{
A vector string indicating either that the tibble does not have empty
observation or the row number of the empty observations.
}
\description{
This helper function extracts the row number(s) having NA values for all
columns.
}
\examples{
{

##### Example 1 -------------------------------------------------------------
# All rows have observation
get_all_na_rows2(iris)

##### Example 2 -------------------------------------------------------------
# One row doesn't have any observations
library(dplyr)
get_all_na_rows2(tbl = bind_rows(iris, tibble(Species = c(NA,NA))))

}

}
