% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-functions_template.R
\name{get_args}
\alias{get_args}
\title{xxx xxx xxx}
\usage{
get_args(...)
}
\arguments{
\item{...}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
