% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-check_table_content.R
\name{get_duplicated_cols}
\alias{get_duplicated_cols}
\title{xxx xxx xxx}
\usage{
get_duplicated_cols(tbl)
}
\arguments{
\item{tbl}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
