% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-check_table_values.R
\name{get_duplicated_rows2}
\alias{get_duplicated_rows2}
\title{Extract observations (rows) that have same values in a tibble (no ID)}
\usage{
get_duplicated_rows2(tbl)
}
\arguments{
\item{tbl}{R object(dataframe or tibble) of the input tibble}
}
\value{
A tibble indicating which row which values is the same in the tibble
}
\description{
This helper function extracts the row number (or first column value) in a
tibble having identical values for all columns. This function can be used
either on the whole columns or excluding the first column (id) (which can be
useful to identify repeated observation across different ids)
}
\examples{
{

# the row numbers are returned to identify which observations have repeated
# values
library(dplyr)

get_duplicated_rows2(tbl = bind_rows(mtcars,mtcars[1,]))

}

}
