% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-lazy.R
\name{get_index_list}
\alias{get_index_list}
\title{xxx xxx xxx}
\usage{
get_index_list(.obj, map_list = NULL)
}
\arguments{
\item{.obj}{xxx xxx xxx}

\item{map_list}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
