% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-check_table_content.R
\name{get_unique_value_cols}
\alias{get_unique_value_cols}
\title{Extract columns that are all 'NA' from a dataset}
\usage{
get_unique_value_cols(dataset)
}
\arguments{
\item{dataset}{A character string or tibble of the input dataset}
}
\value{
A vector string indicating either that the dataset does not have empty
columns or the names of the empty columns.
}
\description{
This helper function extracts the names of the columns in a dataset having NA
values for all observations.
}
\examples{
\dontrun{
# Example 1: All columns have observation
get_all_na_cols(iris)

# Example 2: One column doesn't have any observations
get_all_na_cols(iris \%>\% mutate(new_col = NA))
}

}
