% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-lazy.R
\name{make_name_list}
\alias{make_name_list}
\title{xxx xxx xxx}
\usage{
make_name_list(args_list, list_elem)
}
\arguments{
\item{args_list}{xxx xxx xxx}

\item{list_elem}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.
# Example 1: xxx

}

}
