% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-lazy.R
\name{message_on_prompt}
\alias{message_on_prompt}
\title{xxx}
\usage{
message_on_prompt(...)
}
\arguments{
\item{...}{String character to put in a message}
}
\value{
xxx
}
\description{
xxx
}
\examples{
\dontrun{
# example xxx

xxx

}

}
