% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-functions_template.R
\name{prep_template_run}
\alias{prep_template_run}
\title{xxx xxx xxx}
\usage{
prep_template_run(obj)
}
\arguments{
\item{obj}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
