% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{read_csv_any_formats}
\alias{read_csv_any_formats}
\title{Read a csv file using read_csv and avoid errors}
\usage{
read_csv_any_formats(csv_name)
}
\arguments{
\item{csv_name}{A character string of the path of the csv file.}
}
\value{
a tibble corresponding to the csv read.
}
\description{
The csv file is read twice to detect the number of lines to use in attributing
the column type ('guess_max' parameter of read_csv). This avoids common errors
when reading csv files.
}
\examples{
\dontrun{

# Example 1: xxx
}

}
