% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-functions_template.R
\name{set_template}
\alias{set_template}
\title{xxx xxx xxx}
\usage{
set_template(attributes, version = NULL)
}
\arguments{
\item{attributes}{xxx xxx xxx}

\item{version}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
