% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-functions_template.R
\name{show_pipeline}
\alias{show_pipeline}
\title{xxx xxx xxx}
\usage{
show_pipeline(.fargs)
}
\arguments{
\item{.fargs}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{

# Example 1: xxx xxx xxx.


}

}
