% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-viz.R
\name{template_visual_report}
\alias{template_visual_report}
\title{Create a bookdown template for the visual report}
\usage{
template_visual_report(to = getwd())
}
\arguments{
\item{to}{A character string of a path where the bookdown report will be
placed}
}
\value{
A folder containing all files (Rmd, yml, docs, ...) to generate bookdown
report
}
\description{
This helper function creates a template for the visual report bookdown. This
template is taken from the following link:
https://github.com/jtr13/bookdown-template/archive/refs/heads/master.zip
folder
}
\examples{
\dontrun{
# Example 1: create a folder containing template

template_visual_report("template")

}

}
