% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dates.R
\name{which_any_date}
\alias{which_any_date}
\title{xxx xxx xxx}
\usage{
which_any_date(
  x,
  format = c("dmy", "dym", "ymd", "ydm", "mdy", "myd", "as_date")
)
}
\arguments{
\item{x}{object to be coerced.}

\item{format}{object to be coerced.}
}
\value{
xxx xxx xxx
}
\description{
Creates objects of type "Date".
}
\examples{
\dontrun{
# Example 1: This example


}

}
