% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{write_excel_allsheets}
\alias{write_excel_allsheets}
\title{Write all Excel sheets using \code{\link[writexl:write_xlsx]{writexl::write_xlsx()}} recursively}
\usage{
write_excel_allsheets(list, filename)
}
\arguments{
\item{list}{R objects, coma separated.}

\item{filename}{A character string of the path of the Excel file.}
}
\value{
Nothing to be returned. The file is created at the path declared in the
environment.
}
\description{
The R objects are read and the values are placed in separated sheets.
This function is inspired by the function proposed in
https://statmethods.wordpress.com/2014/06/19/quickly-export-multiple-r-objects-to-an-excel-workbook/
}
\examples{
{

unlink(
  write_excel_allsheets(
    list = list(iris = iris, mtcars = mtcars),
    filename = tempfile()))

}

}
\seealso{
\code{\link[writexl:write_xlsx]{writexl::write_xlsx()}}
}
