% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-dates.R
\name{as_any_date}
\alias{as_any_date}
\title{xxx xxx xxx}
\usage{
as_any_date(x, format = c("dmy", "dym", "ymd", "ydm", "mdy", "myd"))
}
\arguments{
\item{x}{object to be coerced.}

\item{format}{object to be coerced.}
}
\value{
xxx xxx xxx
}
\description{
Creates objects of type "Date".
}
\examples{
\dontrun{
# Example 1: This example


}

}
