% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{as_any_symbol}
\alias{as_any_symbol}
\title{Create objects of type "symbol".}
\usage{
as_any_symbol(x)
}
\arguments{
\item{x}{Object to be coerced or tested. Can be a vector, a character string,
a symbol.}
}
\value{
Object of type "symbol".
}
\description{
Create or test for objects of type "symbol".
}
\examples{
\dontrun{

as_any_symbol(coucou)
as_any_symbol("coucou")

}

}
