% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{template_visual_report}
\alias{fabR_help}
\title{Deprecated functions}
\usage{
template_visual_report(...)

fabR_help(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link[=bookdown_template]{bookdown_template()}} instead of \code{template_visual_report()}.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link[=fabR_website]{fabR_website()}} instead of \code{fabR_help()}.
}
\keyword{internal}
