% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dates.R
\name{guess_date_format}
\alias{guess_date_format}
\title{xxx xxx xxx}
\usage{
guess_date_format(tbl, col = NULL)
}
\arguments{
\item{tbl}{xxx}

\item{col}{xxx}
}
\value{
xxx xxx xxx.
}
\description{
xxx xxx xxx.
}
\examples{
\dontrun{
# Example 1: xxx xxx xxx.

}

}
