% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{make_name_list}
\alias{make_name_list}
\title{Shortcut to create beautiful names in a list}
\usage{
make_name_list(args_list, list_elem)
}
\arguments{
\item{args_list}{xxx xxx xxx}

\item{list_elem}{xxx xxx xxx}
}
\value{
xxx xxx xxx.
}
\description{
Generate a name for an element in a list. This function is targeted for functions
creations which handle lists. Those lists may need names to go through each elements.
This function works with [stats::setNames()] and allows the user to provide name
shorter, more user-friendly in their lists.
}
\examples{
\dontrun{

# Example 1: yyy yyy yyy

}

}
