% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-lazy.R
\name{parceval}
\alias{parceval}
\title{xxx}
\usage{
parceval(...)
}
\arguments{
\item{...}{String character to put in a message}
}
\value{
A list of tibbles corresponding to the sheets read, or a single tibble
if the number of sheets is one.
}
\description{
xxx
}
\examples{
\dontrun{
# example xxx

xxx

}

}
