% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-viz.R
\name{plot_date}
\alias{plot_date}
\title{Draw lollipop plot of one (possibly grouped) time-related column in a tbl}
\usage{
plot_date(
  tbl = "airquality",
  col = "Day",
  filter = "c()",
  negate = FALSE,
  missing_values = "c()",
  time = "day",
  out = "ggplot2",
  group_by = NULL
)
}
\arguments{
\item{tbl}{A character string or tibble specifying the input tbl}

\item{col}{A character string specifying a column of interest}

\item{filter}{A character string specifying the values to filter. (equivalent of 'values in')
This determines which values should be retained. It can be applied to both grouped
and ungrouped data.}

\item{negate}{If TRUE, return non-matching elements.}

\item{missing_values}{Vector listing values to exclude from valid values. These
values will not be excluded from counting - but will be displayed separately from valid values.}

\item{time}{parameter following lubridate synthaxe to specify the period of time
to consider. Can be ymd, mdy, year, months, etc. See lubridate documentation.}

\item{out}{parameter that specifies the output expected: can be either
'ggplot2', 'plotly','ggplot2-code', 'plotly-code','ggplot2-cat' or 'plotly-cat'.
gglot2 renders a static plot, plotly a dynamic plot, code gives the code in a string
(usable directly with eval/parse functions) and cat provides indented code in the
console.}

\item{group_by}{A character string of one column in the tbl that can be
taken as a grouping column. The visual element will be grouped and displayed
by this column.}
}
\value{
lollipop plot object
}
\description{
This function draws a lollipop plot of the values of time related column.
the 'time' parameter uses lubridate synthax to specify the period of time to consider.
Missing values can be given as input to non-valid and valid values separately, or
grouped by another column. The output can be editable (using plotly library) or static
(using ggplot2 library). The R-code is also editable for coding recycling purpose.
}
\examples{
\dontrun{
# Example 1: cat output generated as a template when no argument provided
plot_date()

# Example 2: graph of

tbl = read_csv_any_formats("study_TOKYO.csv")
plot_date(tbl, col = "dob", out = "ggplot2",time = "year")
plot_date(tbl, col = "dob", out = "ggplot2",time = "month")

}

}
