% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-lazy.R
\name{write_excel_allsheets}
\alias{write_excel_allsheets}
\title{Write all Excel sheets using xlsx::write.xlsx recursively}
\usage{
write_excel_allsheets(filename, list)
}
\arguments{
\item{filename}{A character string of the path of the Excel file.}

\item{list}{R objects, coma separated.}
}
\description{
The R objects are read and the content is placed in separated sheets.
This function is inspired by the function proposed in
https://statmethods.wordpress.com/2014/06/19/quickly-export-multiple-r-objects-to-an-excel-workbook/
}
\examples{
\dontrun{

# Example 1: xxx
}

}
